bugs fixes,controller setup, postman

+++ src/utils/cloudinary.js
@@ -17,11 +17,15 @@ const uploadOnCloudinary= async (localFilePath)=>{
             resource_type:'auto'
         })
         // file has been successfully uploaded
-        console.log("file has been uploaded successfully: ",response.url)
+        // console.log("file has been uploaded successfully: ",response.url)
+        // console.log("file has been uploaded successfully: ",response)
+        fs.unlinkSync(localFilePath)
         return response;
     
     } catch (error) {
+        if (fs.existsSync(localFilePath)) {
         fs.unlinkSync(localFilePath); //remove the locally saved temporary file as upload operation got failed
+        }
         return null;
     } 
 }
